creature: move Creature_Initialise to trx
